move: isSupportColonyClient to utils
