+++ src/handlers/actions/moveFunds.ts
@@ -1,10 +1,3 @@
-import {
-  AnyColonyClient,
-  ColonyClientV1,
-  ColonyClientV2,
-  ColonyClientV3,
-  ColonyClientV4,
-} from '@colony/colony-js';
 import { BigNumber, utils } from 'ethers';
 
 import {
@@ -18,22 +11,9 @@ import {
   getDomainDatabaseId,
   verbose,
   getCachedColonyClient,
+  isSupportedColonyClient,
 } from '~utils';
 import provider from '~provider';
-
-/**
- * The handler makes use of colonyClient getDomainFromFundingPot method which is only
- * available on ColonyClientV5 and above. The following type predicate allows to check
- * we're dealing with a client that supports this method
- */
-type SupportedColonyClient = Exclude<
-  AnyColonyClient,
-  ColonyClientV1 | ColonyClientV2 | ColonyClientV3 | ColonyClientV4
->;
-const isSupportedColonyClient = (
-  colonyClient: AnyColonyClient,
-): colonyClient is SupportedColonyClient =>
-  (colonyClient as SupportedColonyClient).getDomainFromFundingPot !== undefined;
 
 export default async (event: ContractEvent): Promise<void> => {
   const receipt = await provider.getTransactionReceipt(event.transactionHash);
