+++ src/utils/clients.ts
@@ -1,4 +1,12 @@
-import { AnyVotingReputationClient, Extension } from '@colony/colony-js';
+import {
+  AnyColonyClient,
+  AnyVotingReputationClient,
+  ColonyClientV1,
+  ColonyClientV2,
+  ColonyClientV3,
+  ColonyClientV4,
+  Extension,
+} from '@colony/colony-js';
 import networkClient from '../networkClient';
 
 export const getVotingClient = async (
@@ -7,3 +15,17 @@ export const getVotingClient = async (
   const colonyClient = await networkClient.getColonyClient(colonyAddress);
   return await colonyClient.getExtensionClient(Extension.VotingReputation);
 };
+
+/**
+ * The handler makes use of colonyClient getDomainFromFundingPot method which is only
+ * available on ColonyClientV5 and above. The following type predicate allows to check
+ * we're dealing with a client that supports this method
+ */
+type SupportedColonyClient = Exclude<
+  AnyColonyClient,
+  ColonyClientV1 | ColonyClientV2 | ColonyClientV3 | ColonyClientV4
+>;
+export const isSupportedColonyClient = (
+  colonyClient: AnyColonyClient,
+): colonyClient is SupportedColonyClient =>
+  (colonyClient as SupportedColonyClient).getDomainFromFundingPot !== undefined;
